Move new System.Native shim methods out of pal_environment
pal_environment should only contain methods related to env variables to make the iOS specific variant work well.
