+++ src/native/libs/System.Native/pal_environment.h
@@ -12,9 +12,3 @@ PALEXPORT char** SystemNative_GetEnviron(void);
 
 PALEXPORT void SystemNative_FreeEnviron(char** environ);
 
-PALEXPORT int32_t SystemNative_SchedGetCpu(void);
-
-PALEXPORT __attribute__((noreturn)) void SystemNative_Exit(int32_t exitCode);
-
-PALEXPORT __attribute__((noreturn)) void SystemNative_Abort(void);
-
