+++ src/native/libs/System.Native/pal_threading.h
@@ -23,3 +23,9 @@ PALEXPORT int32_t SystemNative_LowLevelMonitor_TimedWait(LowLevelMonitor *monito
 PALEXPORT void SystemNative_LowLevelMonitor_Signal_Release(LowLevelMonitor* monitor);
 
 PALEXPORT int32_t SystemNative_RuntimeThread_CreateThread(uintptr_t stackSize, void *(*startAddress)(void*), void *parameter);
+
+PALEXPORT int32_t SystemNative_SchedGetCpu(void);
+
+PALEXPORT __attribute__((noreturn)) void SystemNative_Exit(int32_t exitCode);
+
+PALEXPORT __attribute__((noreturn)) void SystemNative_Abort(void);
